# Improve convert_hex_to_binary x86_64 codegen.