@lemire 2023-08-20
When convert_hex_to_binary, we know that the character is an hexadecimal digit so it is a nice function to optimize.


Your new code should sustain a throughput of 1 routine per ~slightly more than 2 cycles.

Can we do better?


What about...

```C++
unsigned convert_hex_to_binary(const char c)  {
  const static unsigned char table[] = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 
                                         0, 0, 0, 0, 0, 0, 0, 10, 11, 
                                         12, 13, 14, 15, 0, 0, 0, 0, 
                                          0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
                                          0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
                                          0, 0, 10, 11, 12, 13, 14, 15};
  return table[c - '0'];
}
```

This should compile to something like...

```asm
        movsxd  rax, edi
        lea     rcx, [rip + table(char)::table]
        movzx   eax, byte ptr [rax + rcx - 48]
        ret
```

It is about and should support a throughput of 1 per cycle, so it is unbeatable from that point of view.

The table fits in a cache line, so it is going to be efficient.