@lemire 2023-08-20
Note that we have as a homework to greatly optimize this work via...

https://github.com/ada-url/ada/pull/459/files

We still don't have good benchmarks for this (at least not benchmarks that are in place).