@lemire 2023-08-21
@ttsugriy  

The function you propose compiles to about 30 bytes (x64). The function I propose compiles to about 16 bytes, plus a table of 54 bytes... so about 70 bytes. Not counting alignment and function headers, it is a difference of about 40 bytes. Of course, it is a short function so it is likely to get inlined. The table will not be duplicated, but the code itself might be. So it is unclear to me which is going to generate a smaller binary. Let us try it out.

I use GCC 11.

Current main branch:
```
$ ls -al build/src/libada.a      
-rw-r--r--  1 dlemire  staff  494544 build/src/libada.a
```

This PR:
```
$ ls -al buildpatch1/src/libada.a
-rw-r--r--  1 dlemire  staff  494544 buildpatch1/src/libada.a
```

PR with a table:
```
$ ls -al buildpatch2/src/libada.a
-rw-r--r--  1 dlemire  staff  494536 buildpatch2/src/libada.a
```

So the function with a table actually saves 8 bytes.

Does it matter? Well. It is 0.0016 % and current processors have megabytes of cache.

Now, we can, of course, measure this empirically, by running benchmarks. We don't yet have good benchmarks for this problem, but I hope to have some later today. So we will be able to measure the difference objectively.