@ttsugriy 2023-08-21
@lemire

I wasn't making any claims about binary size - what I'm saying is that lookup table requires extra memory load and in my experience memory is more often a bottleneck than CPU, so in practice I usually observe runtime reduction in case a few extra CPU instructions can replace a memory access. Obviously only benchmark using representative workload can answer this, but in any case this PR seems like an improvement over original implementation whereas change to lookup table, while may be even better, is much harder to compare in terms of performance.