@lemire 2023-08-21
I think we need a benchmark before we can consider such a PR.

I have the following for consideration: https://github.com/ada-url/ada/pull/477