@zingaburga 2023-08-21
If you want to avoid branches + lookup table, this might be shorter:

	unsigned convert_hex_to_binary(const char c) noexcept {
	  char c2 = c | 32; // convert to lowercase
	  return (c2>='a') ? c2-'a'+10 : c2-'0';
	}

x86-64 codegen:

        or      dil, 32
        movsx   ecx, dil
        lea     edx, [rcx - 48]
        lea     eax, [rcx - 87]
        cmp     cl, 97
        cmovl   eax, edx
        ret

ARM64 codegen:

        orr     w9, w0, #0x20
        mov     w8, #-48
        and     w9, w9, #0xff
        mov     w10, #-87
        cmp     w9, #96
        csel    w8, w10, w8, hi
        add     w0, w8, w9
        ret