@zingaburga 2023-08-22
Gah, scrap that, found a shorter version:

	unsigned convert_hex_to_binary(const char c) {
	  char c2 = c & ~48; // convert to uppercase + set '0' to 0
	  return (c2>'A'-10) ? c2-'A'+10 : c2;
	}

x86-64:

        and     dil, -49
        movsx   ecx, dil
        lea     eax, [rcx - 55]
        cmp     cl, 56
        cmovl   eax, ecx
        ret

ARM64:

        mov     w8, #207
        and     w8, w0, w8
        subs    w9, w8, #55
        csel    w0, w9, w8, hi
        ret