@lemire 2023-08-23
@zingaburga  Very nice. Very slightly slower than a table in my tests, but seemingly faster than this PR.