@zingaburga 2023-08-23
Thanks!

On x86, it's +2 instructions to avoid a lookup, but on ARM64, it should roughly be the same.

Actually, it should be possible to eliminate the sign extension, but I can't seem to get the compiler to comply.  [This](https://godbolt.org/z/o3K1rn1j6) seems to work (unless I misunderstand something), but don't know how to get the compiler to produce it naturally.  This would reduce it to 4 instructions, or effectively 3 with move elimination, which should put it almost on par with the lookup table (but without a lookup).